fix mast module path
