+++ air/src/lib.rs
@@ -9,9 +9,8 @@ extern crate std;
 use alloc::vec::Vec;
 
 use vm_core::{
-    mast::ProgramInfo,
     utils::{ByteReader, ByteWriter, Deserializable, Serializable},
-    ExtensionOf, StackInputs, StackOutputs, ONE, ZERO,
+    ExtensionOf, ProgramInfo, StackInputs, StackOutputs, ONE, ZERO,
 };
 use winter_air::{
     Air, AirContext, Assertion, EvaluationFrame, ProofOptions as WinterProofOptions, TraceInfo,
