+++ assembly/src/assembler/mod.rs
@@ -10,8 +10,8 @@ use crate::{
 };
 use alloc::{boxed::Box, sync::Arc, vec::Vec};
 use vm_core::{
-    mast::{Kernel, MastForest, MastNode, MastNodeId, MerkleTreeNode},
-    Decorator, DecoratorList, Operation,
+    mast::{MastForest, MastNode, MastNodeId, MerkleTreeNode},
+    Decorator, DecoratorList, Kernel, Operation,
 };
 
 mod basic_block_builder;
