+++ core/src/kernel.rs
