+++ miden/tests/integration/operations/io_ops/env_ops.rs
@@ -1,6 +1,9 @@
 use processor::FMP_MIN;
 use test_utils::{build_op_test, build_test, StackInputs, Test, Word, STACK_TOP_SIZE};
-use vm_core::{MastForest, MastNode, MerkleTreeNode, Operation};
+use vm_core::{
+    mast::{MastForest, MastNode, MerkleTreeNode},
+    Operation,
+};
 
 // SDEPTH INSTRUCTION
 // ================================================================================================
