+++ processor/src/decoder/tests.rs
@@ -18,8 +18,8 @@ use miden_air::trace::{
 };
 use test_utils::rand::rand_value;
 use vm_core::{
-    mast::{BasicBlockNode, Program, OP_BATCH_SIZE},
-    MastForest, MastNode, MerkleTreeNode, EMPTY_WORD, ONE, ZERO,
+    mast::{BasicBlockNode, MastForest, MastNode, MerkleTreeNode, OP_BATCH_SIZE},
+    Program, EMPTY_WORD, ONE, ZERO,
 };
 
 // CONSTANTS
