+++ verifier/src/lib.rs
@@ -17,11 +17,7 @@ use winter_verifier::verify as verify_proof;
 // EXPORTS
 // ================================================================================================
 
-pub use vm_core::{
-    chiplets::hasher::Digest,
-    mast::{Kernel, ProgramInfo},
-    StackInputs, StackOutputs, Word,
-};
+pub use vm_core::{chiplets::hasher::Digest, Kernel, ProgramInfo, StackInputs, StackOutputs, Word};
 pub use winter_verifier::{AcceptableOptions, VerifierError};
 pub mod math {
     pub use vm_core::{Felt, FieldElement, StarkField};
